Take load off non-MUC domains
We now scan non-MUC domains only every (approx.) six hours. This
is achieved by looking at the identity of the domain to identify
it as MUC and the last_seen timestamp on each scan.

If the domain is a MUC, it is included in all scans.

If the domain is not a MUC, it is only included if it never had a
successful scan or if the last successful scan has been more than
six hours ago. This is well before the stale state of a domain,
and also if the scan after six hours fails, the next scan will
again include the domain (since the condition still holds true).
This means after dropping out, a domain has at least 18 hours
to be fixed before it is marked as stale.