Display number of users below avatar
This works on all screen types immediately and looks very nice.
Had to adapt the CSS a little to keep the line rhythm. Now the
avatar has the height of exactly two lines of text/address in the
card, and the number of users is at the same height as the third
line of text (if it exists).