Set Expires header on badges
The Expires header is set to the timestamp of the last update plus
one hour. This is consistent with the crawl time. Each crawl will
update the MUC due to the update of the moving average, so this
makes sense.

In some cases due to shuffling of the MUC items in a crawl it is
possible that a badge is updated with a delay, but I think people
are used to web caches and their effects enough for this to not
matter.