Set Expires header on avatars
For avatars, we set the Expires header to the current time plus
one hour. This is consistent with the crawl interval. Since avatars
are not updated on each crawl, the last_updated value (even plus
one hour) will be very much in the past, so we cannot use this to
reduce staleness of avatars on some clients.