Massively improve response times of db-heavy views
Apparently, constructing SQLalchemy mapped objects can be veeery
slow. I’ve seen differences of a factor of 5 between querying
db.session.query(model.MUC, model.PubliclyListedMUC) and the
new {view,api}_base_query functions.

The result is a bit more unwieldly to work with, but the changes
especially in the template were surprisingly tame.

So let’s go for this. I expect the load on the service to drop
considerably.